requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
gspread>=6.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
bcrypt>=4.1.2
groq>=0.1.0
plotly>=5.18.0
